LevelSetEvolution

In GridapTopOpt, the level-set function is evolved and reinitialised using implementations of LevelSetEvolution. We provide the following implementations of this:

Finite differences

Unfitted finite element methods

Custom LevelSetEvolution

To implement a custom level-set evolution method, the following methods can be extended. For example, one could consider Reaction-Diffusion-based evolution of the level set function. This can be solved with a finite element method and so we can implement a new type that inherits from LevelSetEvolution independently of the Stencil types.

GridapTopOpt.LevelSetEvolution — Type

abstract type LevelSetEvolution

Your own evolution method can be implemented by implementing concrete functionality of the below.

GridapTopOpt.evolve! — Method

evolve!(::LevelSetEvolution,φ,args...)

Evolve the level set function φ according to an evolution method LevelSetEvolution.

GridapTopOpt.reinit! — Method

reinit!(::LevelSetEvolution,φ,args...)

Reinitialise the level set function φ according to an evolution method LevelSetEvolution.

GridapTopOpt.get_dof_Δ — Method

get_dof_Δ(::LevelSetEvolution)

Return the distance betweem degree of freedom